Web/bank_simulator/requirements.txt
Flask==3.1.3
Flask-MySQLdb ==2.0.0
python-dotenv